--- a/.prettierignore
+++ b/.prettierignore
@@ -1,0 +1,5 @@
+node_modules
+.next
+out/
+public
+pnpm-lock.yaml