--- a/src/components/Buttons/index.ts
+++ b/src/components/Buttons/index.ts
@@ -1,0 +1,2 @@
+export { default as Button } from './Button'
+export { default as ShareButton } from './ShareButton'