--- a/src/components/Main/index.ts
+++ b/src/components/Main/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Main'