--- a/src/utils/getIconsCount.ts
+++ b/src/utils/getIconsCount.ts
@@ -1,0 +1,4 @@
+import { importIcons } from './import-icons'
+
+export const getIconsCount = () =>
+  importIcons(require.context(`heroicons/24/outline/`, false, /\.svg$/)).length