… .eslintrc.json
… .github/dependabot.yml
… README.md
… jsconfig.json
… package.json
… pnpm-lock.yaml
… public/apple-touch-icon.png
… public/beams.jpg
… public/favicon-16x16.png
… public/favicon-32x32.png
… public/social-card.jpg
… src/components/Main/SearchBar.tsx
… src/components/UI/Icons/index.tsx
… src/components/UI/Illustration.tsx
… src/fonts/Inter-italic.var.woff2
… src/fonts/Inter-roman.var.woff2
… src/pages/_app.tsx
… src/pages/_document.js
… src/pages/index.js
… src/styles/tailwind.css
… src/utils/import-icons.ts
… tailwind.config.js
… tsconfig.json
… v1/next.config.js
… v1/src/data/tags.js
… v1/src/pages/index.js