### Pain Points

-  Everything on the single page, not using components.
-  Not using Typescript.
-  Lighthouse score on mobile is in ~70s.
-  TBT (total blocking time) is huge and a minor CLS from font change.
-  Using deprecated webpack feature.
-  Using Tabs from `headless UI` results in re-mounting of Icons which is expensive and slow.

### What I have done

-  Started with a new `create-next-app`.
-  Updated dependencies to latest stable.
-  Fixed Hydration issues (client server mismatch)
-  Reduced code repetition, Extracted components.
-  Added Suspense boundaries.
-  Enable Browser List For SWC experimental feature.
-  Used `asset/source` from  webpack5 instead of `raw-loader`.
-  Used `fontsource` for Inter font (reducing page size by ~186 KB).
-  Moved Images to public dir for caching.
-  Exactly same layout with one exception, letter 'a' in font is different.
-  Used opinionated file structure.
-  Used `menu` from `headless UI` to select the desired icon group
-  Moved some dependencies to devDependencies.
-  Configured ESlint with recommended Next.js rules.
-  Fixed some ESlint errors.
-  Added missing HTML lang attribute.
-  Ran prettier.
-  Used Semantic HTML tags.

## Lighthouse Scores

|   |  Before |  After  |  
|---|---|---|
|  Mobile  | ![Screenshot 2022-09-06 at 01-31-04 Performance Report for https __heroicons com_ SpeedVitals](https://user-images.githubusercontent.com/80816622/188510585-3bc94da8-306b-4709-a5fe-ecf83538f55f.png)  | ![Screenshot 2022-09-06 at 01-31-35 Performance Report for https __heroicons-optimized vercel app_ SpeedVitals](https://user-images.githubusercontent.com/80816622/188510600-0eccc02c-64db-4cb1-9144-5333eb320502.png)  |  
|  Desktop | ![Screenshot 2022-09-06 at 01-31-27 Performance Report for https __heroicons com_ SpeedVitals](https://user-images.githubusercontent.com/80816622/188510663-0742971c-4ae6-4f8e-aa25-0829b22b2a45.png)  | ![Screenshot 2022-09-06 at 01-31-41 Performance Report for https __heroicons-optimized vercel app_ SpeedVitals](https://user-images.githubusercontent.com/80816622/188510680-e0e04c08-71e8-4036-9998-38dc4bb988db.png)  |



![Screenshot 2022-09-04 at 02-20-07 PageSpeed Compare for Easy Page Speed Benchmarking](https://user-images.githubusercontent.com/80816622/188511168-03b2a6ee-bd6c-44a5-af72-3acbbfeb32c0.png)

![Screenshot 2022-09-04 at 02-19-58 PageSpeed Compare for Easy Page Speed Benchmarking](https://user-images.githubusercontent.com/80816622/188511162-bcf0b020-8df0-42ad-a075-1442ae3b6a9f.png)


Some tests:
https://gtmetrix.com/compare/uyjmc0KU/V0bdBX9u
https://www.webpagetest.org/video/compare.php?tests=220901_BiDcG0_FE1,220901_BiDcXA_FE2
https://www.webpagetest.org/video/compare.php?tests=220902_BiDcS4_4MF,220902_BiDc6S_4MG
https://www.webpagetest.org/video/compare.php?tests=220902_BiDcVM_4PB,220902_AiDcAZ_4M9

Live URL https://heroicons-optimized.vercel.app/


Thanks.