@adamwathan 2022-09-06
Hey! Wow, I appreciate the effort here but unfortunately this is just way too big of a PR for us to ever find time to reasonable review and merge 😕 Maintaining this website isn't a source of pain for us at all so it's not a high priority for us to make any changes to the actual structure of the project.

It looks like there are a handful of things in here that would be good improvements for sure (like the missing `lang` attribute) but we can't really review or merge this at this size, just have too many existing priorities to schedule the time to go through it. I don't mind accepting separate PRs that are much smaller in scope for individual improvements but candidly we mostly dread getting PRs unless they are either extremely small and a no-brainer to merge — we just have limited time and need to spend it on things we're prioritizing ourselves.

Again I appreciate all the work that went into it, hopefully you can understand though why it's hard to fit this into our schedule ❤️